// src/app/models/creditos-api-response.ts
export interface CreditosApiResponse {
  id: number;
  cast: MembroElenco[];
  crew: MembroProducao[];
}

export interface MembroElenco {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string;
  cast_id: number;
  character: string;
  credit_id: string;
  order: number;
}

export interface MembroProducao {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string;
  credit_id: string;
  department: string;
  job: string;
}